# src/py/scm/generated/v1/resources/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "supply-chain-monkey",
    "version": "2026.8.12",
    "description": "Public HTTP and JSON contract for Supply Chain Monkey v1.\n\nSCM normalizes electronic-component data from configured supplier backends.\nExcept for health and the HTML status page, operations require a bearer token."
  },
...
  "paths": {
...
    "/v1/providers/status": {
      "get": {
        "operationId": "AuthenticatedOperations_providersStatus",
        "summary": "List provider status and capabilities",
        "description": "Return configuration and capabilities for every known provider.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Response returned when bearer authentication fails.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Response returned when request validation fails.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorDetail"
                }
              }
            }
          },
          "500": {
            "description": "Response returned when SCM authentication is not configured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HttpErrorDetail": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable request or server error."
          }
        },
        "unevaluatedProperties": {
          "not": {}
        },
        "description": "FastAPI-compatible HTTP error body."
      },
      "JsonInteger": {
        "type": "integer",
        "format": "int64",
        "minimum": -9007199254740991,
        "maximum": 9007199254740991,
        "description": "I-JSON interoperable integer range used by SCM JSON payloads."
      },
      "JsonValue": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "type": "null"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JsonValue"
            }
          },
          {
            "type": "object",
            "unevaluatedProperties": {
              "$ref": "#/components/schemas/JsonValue"
            }
          }
        ],
        "description": "JSON-compatible scalar, array, or object value."
      },
...
      "ProviderStatusEntry": {
        "type": "object",
        "required": [
          "configured",
          "capabilities"
        ],
        "properties": {
          "configured": {
            "type": "boolean",
            "description": "Whether the provider is currently configured for use."
          },
          "backend": {
            "type": "string",
            "description": "Selected backend implementation, when available."
          },
          "capabilities": {
            "$ref": "#/components/schemas/SupplierCapabilities",
            "description": "Features and operational limits exposed by the provider."
          }
        },
        "unevaluatedProperties": {
          "not": {}
        },
        "description": "Configuration and capabilities reported for one provider."
      },
      "ProviderStatusResponse": {
        "type": "object",
        "required": [
          "providers"
        ],
        "properties": {
          "providers": {
            "type": "object",
            "unevaluatedProperties": {
              "$ref": "#/components/schemas/ProviderStatusEntry"
            },
            "description": "Provider status keyed by canonical provider name."
          }
        },
        "unevaluatedProperties": {
          "not": {}
        },
        "description": "Map of provider names to current configuration and capabilities."
      },
...
      "SupplierCapabilities": {
        "type": "object",
        "required": [
          "supplier",
          "provider_kind",
          "supports_mpn_search",
          "supports_keyword_search",
          "supports_spn_lookup",
          "supports_native_spn_batch",
          "max_spn_batch_size",
          "min_request_interval_seconds",
          "rate_limit_per_minute",
          "rate_limit_per_day",
          "usage_unit",
          "supports_quota_headers",
          "notes"
        ],
        "properties": {
          "supplier": {
            "type": "string",
            "description": "Canonical supplier represented by the backend."
          },
          "provider_kind": {
            "type": "string",
            "description": "Implementation category used for provider diagnostics."
          },
          "supports_mpn_search": {
            "type": "boolean",
            "description": "Whether manufacturer-part-number search is supported."
          },
          "supports_keyword_search": {
            "type": "boolean",
            "description": "Whether general keyword search is supported."
          },
          "supports_spn_lookup": {
            "type": "boolean",
            "description": "Whether exact supplier-part-number lookup is supported."
          },
          "supports_native_spn_batch": {
            "type": "boolean",
            "description": "Whether the upstream provider has a native batch lookup."
          },
          "max_spn_batch_size": {
            "$ref": "#/components/schemas/JsonInteger",
            "description": "Maximum batch size accepted by SCM for this provider."
          },
          "min_request_interval_seconds": {
            "type": "number",
            "format": "double",
            "description": "Minimum delay SCM applies between provider requests."
          },
          "rate_limit_per_minute": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonInteger"
              },
              {
                "type": "null"
              }
            ],
            "description": "Provider request quota per minute, when known."
          },
          "rate_limit_per_day": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonInteger"
              },
              {
                "type": "null"
              }
            ],
            "description": "Provider request quota per day, when known."
          },
          "usage_unit": {
            "type": "string",
            "description": "Unit consumed by one reported quota usage event."
          },
          "supports_quota_headers": {
            "type": "boolean",
            "description": "Whether upstream quota headers are exposed in envelope metadata."
          },
          "notes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Human-readable capability and operational notes."
          }
        },
        "unevaluatedProperties": {
          "not": {}
        },
        "description": "Search and lookup features exposed by one configured provider backend."
      },
      "ValidationErrorDetail": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            },
            "description": "Individual validation failures."
          }
        },
        "unevaluatedProperties": {
          "not": {}
        },
        "description": "FastAPI-compatible collection of request validation failures."
      },
      "ValidationErrorItem": {
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "properties": {
          "loc": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "$ref": "#/components/schemas/JsonInteger"
                }
              ]
            },
            "description": "Request location of the invalid value."
          },
          "msg": {
            "type": "string",
            "description": "Human-readable validation message."
          },
          "type": {
            "type": "string",
            "description": "Stable validation error type."
          },
          "input": {
            "$ref": "#/components/schemas/JsonValue",
            "description": "Rejected input value, when included by FastAPI."
          },
          "ctx": {
            "type": "object",
            "unevaluatedProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "description": "Additional validation context, when included by FastAPI."
          }
        },
        "unevaluatedProperties": {
          "not": {}
        },
        "description": "One FastAPI-compatible validation failure."
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "Bearer"
      },
      "LegacyQueryTokenAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "Deprecated v1 query-token compatibility surface. Never place a real service token in this query, Swagger operation, logs, or shared URLs."
      }
    }
  }
}